move lib/term.js to src/term.js.
