+++ .gitignore
@@ -7,3 +7,4 @@ Makefile.gyp
 *.Makefile
 *.target.gyp.mk
 *.node
+example/*.log
