+++ lib/index.js
@@ -31,7 +31,7 @@ term.middleware = function(options) {
   };
 };
 
-term.path = __dirname + '/term.js';
+term.path = __dirname + '/../src/term.js';
 
 term.__defineGetter__('script', function() {
   if (term._script) return term._script;
@@ -41,7 +41,7 @@ term.__defineGetter__('script', function() {
 
 term.__defineGetter__('Terminal', function() {
   if (term._Terminal) return term._Terminal;
-  return term._Terminal = require('./term');
+  return term._Terminal = require('../src/term');
 });
 
 /**
